Add ramlog support
When the muc-loader boots to flash mode, add print with reason
to Nuttx ramlog (if ramlog valid). Initially, this will aid
debugging an issue where it is believed the MuC is rebooting
into the muc-loader, but could be expanded for other uses in
the future.

Only this one print is added to the ramlog to avoid a scenario
where the muc-loader prints flood out the Nuttx prints.

Change-Id: I4ae633ece85abcff5c16658fe0f5ac038438b77b
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/845579
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>