configs: hdk: Only attempt to read eeprom if pcard is present
When a pcard is removed, the pcard detection interrupt fires
before the pins on the 80-pin connector are disconnected. Therefore,
when Nuttx resets the MuC, the muc-loader could still read the pcard
eeprom and boot back into Nuttx before the user could fully remove
the pcard. So instead of getting the blinky (the default firmware
when no pcard is attached), the HDK would still report as the
device of the previously attached pcard.

To prevent this, only attempt to read the eeprom of the pcard if
the pcard detect pin indicates a pcard is present. With this change,
when a user removes a pcard, they will actually see the HDK reset
to the blinky target.

Change-Id: Ie8e6879af812e2832babbc6202be2045370592c8
Signed-off-by: Greg Meiste <w30289@motorola.com>
Reviewed-on: http://gerrit.mot.com/872904
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Makarand Karvekar <mkarvekar@motorola.com>
Reviewed-by: James Wylder <jwylder@motorola.com>