gbfirmware: Set flashing flag for any stage
The flashing flag is required to detect an interrupted
flashing.  The flag should be set before any firmware
is modified and cleared after all changes are complete.
Because of the special flow for stage two (the APBE flashing)
the flag was not getting set.  If the update is interrupted
during stage two, the bootloader would not detect this on
the next restart.

Move the writing of the flashing flag to a common section
of code so it is set in all cases.

Change-Id: Ic89d017bc9a5c7af1139202ff1cefa90dda1ab2b
Signed-off-by: Jim Wylder <jwylder@motorola.com>
Reviewed-on: http://gerrit.mot.com/874829
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Reviewed-by: Gregory Meiste <w30289@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>