Use Generic Pin Label for GPIO_PIN_WAKE_N
The base code was not updated to use the generic macro
for the wake signal.  It is on GPIO_PIN_0 on all other
devices so this is the first device to notice.

CRs-fixed: IKVALENCE-1056
Change-Id: If866ec1ef2a221470c9c2a6d746cf63b27f4a987
Signed-off-by: Jim Wylder <jwylder@motorola.com>
Reviewed-on: https://gerrit.mot.com/947310
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Weiqi Tang <woodtang@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>